fix: Fix API wrapper builds and sanitizer issues
- webgpu_wrapper.c: Remove mvgal_initialized, fix submit_webgpu_workload calls, fix enum names
- d3d_wrapper.c: Remove 3rd arg from all submit_d3d_workload calls
- metal_wrapper.c: Remove mvgal_initialized references
- CMakeLists.txt: Remove duplicate mvgal SHARED lib, fix sanitizer logic
- runtime/CMakeLists.txt: Only enable sanitizers when MVGAL_ENABLE_SANITIZERS=ON
- Fix E_FAIL sign conversion warning
- Remove unused variables from wrapper functions
- Enable MVGAL_BUILD_API=ON for mvgal_core static library build

All 7 tests passing, all targets building successfully.

+++ src/userspace/memory/dmabuf.c
@@ -76,7 +76,7 @@ mvgal_error_t mvgal_dmabuf_allocate(
         }
         
         if (ioctl(state->system_heap.fd, DMA_HEAP_IOCTL_CASE, &alloc) == 0) {
-            fd = alloc.fd;
+            fd = (int)alloc.fd;
             MVGAL_LOG_DEBUG("DMA-BUF heap allocation: size=%zu, fd=%d", size, fd);
         }
     }
